Chore: added add token modal
